requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
praw==7.5.0
python-dotenv==0.20.0
typing